PM: earlysuspend: Add console switch when user requested sleep state changes.
When early-suspend is enabled, the framebuffer is turned off at
early-suspend. If CONSOLE_EARLYSUSPEND is enabled, a console switch
is used to notify user space that it should stop drawing before
the framebuffer is disabled, and that it should repaint and resume
drawing after the framebuffer is turned back on.

Signed-off-by: Arve Hjønnevåg <arve@android.com>