[ARM] msm: msm_serial_hs: Add optional low power wakeup.
Low power wakeup is configured using msm_serial_hs_platform_data. This allows
us to clk_enable() the uart on another interrupt (for example a gpio rx line).
It also allows us to inject an char into the rx tty layer when this occurs.

This is useful for uart attached devices that have low power modes independant
of the linux CPU, for example bluetooth chip.

Currently the interface to control low power mode is through
msm_hs_request_clock_on()
msm_hs_request_clock_off()
Ideally we want to use the tty/serial layer but we do not have a good solution
to this yet.

A few other misc changes in here
- Use _locked convention for functions that hold spinlock
- request_irq() and free_irq() in startup/shutdown not probe/release
- Use BIT(x) helper in the hw_reg header. Add missing bitmasks.

Signed-off-by: Nick Pelly <npelly@google.com>