msm-nand: Add bad block marking/checking support.
Add support to check if a block is bad by reading the bad block
marker. For marking a block as bad, write all 0's to the first
page of the block which results in the bad block marker being set
to 0 as well.

Acked-by: David Brown <davidb@qualcomm.com>

Removed config option since some factory badblocks are marked in a way
that does not prevent erase.
Signed-off-by: Arve Hjønnevåg <arve@android.com>