binder: Keep a reference to the files_struct while the driver is mmapped
This prevents breaking fget_light if a single threaded application allows
incoming file descriptors (in replies or on nodes).
Should also prevent inserting a file in the wrong files_struct if the
receving process execs in the middle of a transaction (between
task_get_unused_fd_flags and task_fd_install).

Signed-off-by: Arve Hjønnevåg <arve@android.com>