[ARM] msm: driver for MSM7X00A on-chip I2C controller.
San: Re-write driver to write/read to/from buffer in IRQ context

Signed-off-by: Arve Hjønnevåg <arve@android.com>
Signed-off-by: San Mehat <san@android.com>

[ARM] msm: I2C: Send stop command to finish i2c transfer on timeout

Signed-off-by: San Mehat <san@android.com>

[ARM] msm: i2c: Try to read the requested number of bytes.

If the requested read length is two or greater, this should always work.
It the read length is one, an extra byte may be read if the interrupt
handler is not called in time. The previous version always read an extra
byte.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Try to recover if the i2c bus is busy after a failed transfer.

If the bus is busy, switch the i2c pins to gpio mode and drive the bus
manually. This will clear the busy condition with some clients.

This adds a function, msm_set_i2c_mux, to change i2c mux settings since we do
not have mux api.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Perform entire i2c transfer at interrupt context.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Prevent infinite reads.

The contoller sometimes ignores our stop request. Send a new stop request
before flushing every other byte.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Improve error recovery

- Reduce bus busy timeout.
- Call recovery function if the bus is busy at the start of a transaction.
- Wait for bus busy to clear after a transaction.
- Move read stop on timeout to recover function.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Always read at least two bytes

If we tell the controller to stop at the same time as it finished reading
the first byte, it may ignore the stop request, loose bus access or fail
to start the next read or write. We now always read an extra byte if a
single byte read was requested.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Only send STOP for last for the last write message in the group.

This allows drivers to perform a write followed by a read with a repeated
start. A repeated start after a read is still not possible.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Send another byte if we are still the bus master in msm_i2c_recover_bus_busy.

If a transfer timed out before we told the controller to stop writing the
controller may be waiting for more data to write. Write 0xff as the last
byte, after switching the lines to gpio mode, to trigger a failure.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Work around i2c controller bug.

When writing I2C_WRITE_DATA right before the i2c clock line is scheduled
to go high again, the controller does not stretch the clock so the data
setup time is within the i2c spec.

Change-Id: Ibfa35a036454ac39a8bb94b0d1fc220976ace88f
Signed-off-by: Arve Hjønnevåg <arve@android.com>

msm: i2c: only complain if the bus is busy at the start of a transaction

Signed-off-by: Brian Swetland <swetland@google.com>

[ARM] msm: i2c: do the pin mux config at initialization time.

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] msm: i2c: dump the values of scl/sda at the start of recover_bus_busy

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] msm: i2c: Only enable clock during transfer.

The arm9 no longer ignores our clock requests when we enter sleep mode.
Always leaving the i2c clock on will prevent entering low power mode.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

[ARM] msm: i2c: Don't depend on i2c interrupt wakeup.

The other core may use the i2c controller while we are power collpased
which makes the i2c interrupt unusable as a wakeup source.
Hold a wakelock while i2c is active and disable the interrupt when it is idle.

Signed-off-by: Arve Hjønnevåg <arve@android.com>